+++ tests/gmprocess/io/bhrc/bhrc_test.py
@@ -1,5 +1,5 @@
 from gmprocess.io.bhrc.core import is_bhrc, read_bhrc
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 
 
 def test_bhrc():