+++ tests/gmprocess/io/cwb/cwb_test.py
@@ -8,7 +8,7 @@ import numpy as np
 
 # local imports
 from gmprocess.io.cwb.core import is_cwb, read_cwb, _get_header_info
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 
 
 def test():