+++ tests/gmprocess/io/obspy/obspy_test.py
@@ -1,5 +1,5 @@
 from gmprocess.io.obspy.core import read_obspy
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 from gmprocess.core.streamcollection import StreamCollection
 from gmprocess.waveform_processing.processing import process_streams
 from gmprocess.utils.config import get_config