+++ tests/gmprocess/io/obspy/stastionxml_test.py
@@ -4,7 +4,7 @@ from numpy.testing import assert_allclose
 
 from gmprocess.waveform_processing.processing import process_streams
 from gmprocess.core.streamcollection import StreamCollection
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 
 
 def test():