+++ tests/gmprocess/io/renadic/renadic_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from gmprocess.io.renadic.core import is_renadic, read_renadic
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 
 
 def test_renadic():