+++ tests/gmprocess/io/unam/unam_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from gmprocess.io.unam.core import is_unam, read_unam
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 from gmprocess.io.read import read_data