+++ tests/gmprocess/waveform_processing/adjust_highpass_ridder_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 from gmprocess.core.streamcollection import StreamCollection
 from gmprocess.io.read import read_data
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 from gmprocess.waveform_processing.adjust_highpass_ridder import ridder_fchp
 from gmprocess.utils.config import get_config