+++ tests/gmprocess/waveform_processing/corner_frequencies_test.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from gmprocess.core.streamcollection import StreamCollection
 from gmprocess.io.read import read_data
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 from gmprocess.utils.config import get_config
 
 from gmprocess.waveform_processing.windows import signal_split