+++ tests/gmprocess/waveform_processing/processing_test.py
@@ -6,7 +6,7 @@ from gmprocess.core.streamcollection import StreamCollection
 from gmprocess.io.read import read_data
 from gmprocess.waveform_processing.processing import process_streams
 from gmprocess.utils.logging import setup_logger
-from gmprocess.utils.test_utils import read_data_dir
+from gmprocess.utils.tests_utils import read_data_dir
 from gmprocess.utils.config import get_config, update_dict, update_config
 from gmprocess.utils.constants import TEST_DATA_DIR